Puts

+++ pointers_arrays_strings/2-strlen.c
@@ -16,8 +16,7 @@ int _strlen(const char *s)
 
 	{
 	
-		
-	size_t length = 0;
+	int length = 0;
 
 	while (*s++)
 		length++;
